Add DirectoryHTTPHandler
Similar to a FileHTTPHandler, the DirectoryHTTPHandler reads directly serves the files it can read from the file system. However, it finds them by looking inside a specific sub directory for that file.

DirectoryHTTPHandler’s have a ‘server path’: the path to where they will be hosted on the server. This is how it knows where to look on the file system: it removes the server path from the request. Then it uses that subpath, plus its root path, to know where to look.

It uses the FileHTTPHandler’s method of getting MIME types.