Tests for DirectoryHTTPHandler
We’ve added two tests for the DirectoryHTTPHandler.

This required us to make a change to the way SPM embeds the test resources for the test target. Previously, the `.process` method put them all in the top level directory. That’s not handy for our DirectoryHTTPHandler; so we have changed that to `.copy` which preserves the directory structure, and allows us to test better.

This change also meant we needed to update the older tests to specify where they should be looking for their stubs as well.