Support for multihoming with unbound sockets
Multihomed servers (i.e. those with multiple network interfaces) need to take
care to send UDP responses from the interface on which they were received,
otherwise the client may never receive the packet.

Historically this was done by binding separate sockets to each interface,
however this does not adapt well to dynamic interface changes (without extra
code to monitor for this, which is impossible to do in a portable manner).

Modern operating systems provide `IP_PKTINFO` and `IPV6_PKTINFO` which allow
the local interface index and address to be set and reported on unbound
sockets. This commit adds support for this.

Note: currently unavailable on Windows.